+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## 0.14.4
 
+* `Common`:
+    * `JVM`:
+        * New extension `downloadToTempFile`
 * `Ktor`:
     * `Server`:
         * Small fix in `handleUniUpload`
         * `ApplicationCall#uniloadMultipartFile` now uses `uniloadMultipart`
+    * `Common`:
+        * New extension `downloadToTempFile`
     * `Client`:
         * New extensions on top of `uniUpload`